Refactor user caching and notification processing
- Extract common user caching logic to `common.ts`
- Add `clearUsersCache()` and `getAndCacheRecipients()` functions
- Remove duplicate code from `process-weekly-notifications.ts`
- Simplify notification processing scripts
