+++ site/gatsby-site/server/fields/common.ts
@@ -87,6 +87,38 @@ export interface UserAdminData {
     userId?: string;
 }
 
+const usersCache: UserAdminData[] = [];
+
+export const clearUsersCache = () => {
+  usersCache.length = 0;
+}
+
+export const getAndCacheRecipients = async (userIds: string[], context: Context) => {
+
+  const recipients = [];
+
+  for (const userId of userIds) {
+
+    let user = usersCache.find((user) => user.userId === userId) ?? null;
+
+    if (!user) {
+
+      user = await getUserAdminData(userId, context) ?? null;
+
+      if (user) {
+
+        usersCache.push(user);
+      }
+    }
+
+    if (user?.email && user?.userId) {
+      recipients.push({ email: user.email, userId: user.userId });
+    }
+  }
+
+  return recipients;
+}
+
 export const getUserAdminData = async (userId: string, context: Context): Promise<UserAdminData | null> => {
 
     const authUsersCollection = context.client.db('auth').collection("users");
